+++ tests/testthat/test-esr.R
@@ -30,6 +30,33 @@ test_that("esr.mcmcmarray", {
   expect_identical(esr(subset(mcmcr_example[[2]], iterations = 1L), by = "term", as_df = TRUE), tibble(term = as.term(c("parameter[1,1]", "parameter[2,1]", "parameter[1,2]", "parameter[2,2]")), esr = rep(1, 4)))
 })
 
+test_that("esr.mcarray", {
+  mcarray <- as.mcarray(mcmcr::mcmcr_example[[2]])
+  expect_identical(esr(mcarray), 0.05)
+  expect_identical(esr(mcarray, as_df = TRUE), tibble(parameter = "parameter", esr = 0.05))
+  expect_identical(esr(mcarray, by = "parameter"), 0.05)
+  expect_equivalent(esr(mcarray, by = "term"), matrix(c(0.05, 0.05, 0.05, 0.05)))
+  expect_equivalent(esr(mcarray, by = "term", as_df = TRUE), tibble(term = as.term(c("parameter[1,1]", "parameter[2,1]", "parameter[1,2]", "parameter[2,2]")), esr = c(0.05, 0.05, 0.05, 0.05)))
+})
+
+test_that("esr.mcmc", {
+  mcmc <- as.mcmc(subset(mcmcr::mcmcr_example, chains = 2L))
+  expect_identical(esr(mcmc), 0.01)
+  expect_identical(esr(mcmc, as_df = TRUE), tibble(all = "all", esr = 0.01))
+  expect_identical(esr(mcmc, by = "parameter"), list(alpha = 0.01, beta = 0.062, sigma = 0.474))
+  expect_equivalent(esr(mcmc, by = "term"), list(alpha = c(0.01, 0.01), beta = matrix(c(0.062, 0.062, 0.062, 0.062), nrow = 2, ncol = 2), sigma = 0.474))
+  expect_equivalent(esr(mcmc, by = "term", as_df = TRUE), tibble(term = as.term(c("alpha[1]", "alpha[2]", "beta[1,1]", "beta[2,1]", "beta[1,2]", "beta[2,2]", "sigma")), esr = c(0.01, 0.01, 0.062, 0.062, 0.062, 0.062, 0.474)))
+})
+
+test_that("esr.mcmc.list", {
+  expect_identical(esr(as.mcmc.list(mcmcr::mcmcr_example)), 0.011)
+  expect_identical(esr(as.mcmc.list(mcmcr::mcmcr_example), as_df = TRUE), tibble(all = "all", esr = 0.011))
+  expect_identical(esr(as.mcmc.list(mcmcr::mcmcr_example), by = "parameter"), list(alpha = 0.011, beta = 0.05, sigma = 0.472))
+  expect_equivalent(esr(as.mcmc.list(mcmcr::mcmcr_example), by = "term"), list(alpha = c(0.011, 0.011), beta = matrix(c(0.05, 0.05, 0.05, 0.05), nrow = 2, ncol = 2), sigma = 0.472))
+  expect_equivalent(esr(as.mcmc.list(mcmcr::mcmcr_example), by = "term", as_df = TRUE), tibble(term = as.term(c("alpha[1]", "alpha[2]", "beta[1,1]", "beta[2,1]", "beta[1,2]", "beta[2,2]", "sigma")), esr = c(0.011, 0.011, 0.05, 0.05, 0.05, 0.05, 0.472)))
+})
+
+
 test_that("esr.mcmcr", {
   expect_identical(esr(mcmcr_example2), 0.011)
   expect_identical(esr(mcmcr_example), 0.011)